Create FizzBuzz
FizzBuzz is a simple and classic programming game often used to teach basic logic and control flow. The game involves iterating through numbers from 1 up to a specified limit. For each number:

    If the number is divisible by 3, the program outputs "Fizz."
    If the number is divisible by 7, it outputs "Buzz."
    If the number is divisible by both 3 and 7, it outputs "FizzBuzz."
    Otherwise, it simply outputs the number.

An extra element is added in which any number not divisible by 7 or 3, yet had a 3 in it outputs "Almost Fizz".